Moved Metadata Extraction to its own package
